+++ src/AdapterAzureStorageBlob.ts
@@ -53,16 +53,27 @@ export class AdapterAzureStorageBlob extends AbstractAdapter {
         return null;
       }
 
-      const { type, part1: storageAccount, part2: accessKey, bucketName, queryString } = value;
+      const {
+        type,
+        part1: storageAccount,
+        part2: accessKey,
+        bucketName,
+        queryString: options,
+      } = value;
       cfg = {
         type,
         storageAccount,
         accessKey,
         bucketName,
-        ...queryString,
+        ...options,
       };
     } else {
-      cfg = { ...config };
+      if (typeof config.options !== "undefined") {
+        cfg = { ...config, ...config.options };
+        delete cfg.options;
+      } else {
+        cfg = { ...config };
+      }
     }
 
     if (cfg.skipCheck === true) {